+++ Cargo.toml
@@ -15,6 +15,7 @@ thiserror = "1.0.57"
 url = "2.5.0"
 regex = "1.10.3"
 diffy = "0.4.2"
+chrono = { version = "0.4", features = ["serde", "clock"] }
 anyhow = "1.0"
 ortho_config = { git = "https://github.com/leynos/ortho-config", tag = "v0.2.0", default-features = false }
 figment = { version = "0.10", default-features = false, features = ["env", "toml"], optional = true }
