+++ src/main.rs
@@ -4,7 +4,9 @@
 //! summarizes them by file, and prints each thread. When a thread has
 //! multiple comments on the same diff, the diff is displayed only once.
 mod cli_args;
+mod reviews;
 use crate::cli_args::{GlobalArgs, IssueArgs, PrArgs};
+use crate::reviews::{fetch_reviews, latest_reviews, print_reviews};
 use clap::{Parser, Subcommand};
 use figment::error::{Error as FigmentError, Kind as FigmentKind};
 use ortho_config::{OrthoConfig, OrthoError, load_and_merge_subcommand_for};
@@ -263,38 +265,6 @@ struct CommentNode {
     comments: CommentConnection,
 }
 
-#[derive(Debug, Deserialize, Default)]
-struct ReviewConnection {
-    nodes: Vec<PullRequestReview>,
-    #[serde(rename = "pageInfo")]
-    page_info: PageInfo,
-}
-
-#[derive(Debug, Deserialize, Default, Clone)]
-struct PullRequestReview {
-    body: String,
-    #[serde(rename = "submittedAt")]
-    submitted_at: String,
-    state: String,
-    author: Option<User>,
-}
-
-#[derive(Deserialize)]
-struct ReviewData {
-    repository: ReviewRepo,
-}
-
-#[derive(Deserialize)]
-struct ReviewRepo {
-    #[serde(rename = "pullRequest")]
-    pull_request: ReviewPr,
-}
-
-#[derive(Deserialize)]
-struct ReviewPr {
-    reviews: ReviewConnection,
-}
-
 const GITHUB_GRAPHQL_URL: &str = "https://api.github.com/graphql";
 
 /// Width of the line number gutter in diff output
@@ -349,24 +319,6 @@ const COMMENT_QUERY: &str = r"
     }
 ";
 
-const REVIEWS_QUERY: &str = r"
-    query($owner: String!, $name: String!, $number: Int!, $cursor: String) {
-      repository(owner: $owner, name: $name) {
-        pullRequest(number: $number) {
-          reviews(first: 100, after: $cursor) {
-            nodes {
-              body
-              state
-              submittedAt
-              author { login }
-            }
-            pageInfo { hasNextPage endCursor }
-          }
-        }
-      }
-    }
-";
-
 const ISSUE_QUERY: &str = r"
     query($owner: String!, $name: String!, $number: Int!) {
       repository(owner: $owner, name: $name) {
@@ -378,7 +330,7 @@ const ISSUE_QUERY: &str = r"
     }
 ";
 
-async fn paginate<T, F, Fut>(mut fetch: F) -> Result<Vec<T>, VkError>
+pub(crate) async fn paginate<T, F, Fut>(mut fetch: F) -> Result<Vec<T>, VkError>
 where
     F: FnMut(Option<String>) -> Fut,
     Fut: std::future::Future<Output = Result<(Vec<T>, PageInfo), VkError>>,
@@ -447,27 +399,6 @@ async fn fetch_thread_page(
     Ok((conn.nodes, conn.page_info))
 }
 
-async fn fetch_review_page(
-    client: &GraphQLClient,
-    repo: &RepoInfo,
-    number: u64,
-    cursor: Option<String>,
-) -> Result<(Vec<PullRequestReview>, PageInfo), VkError> {
-    let data: ReviewData = client
-        .run_query(
-            REVIEWS_QUERY,
-            json!({
-                "owner": repo.owner.as_str(),
-                "name": repo.name.as_str(),
-                "number": number,
-                "cursor": cursor,
-            }),
-        )
-        .await?;
-    let conn = data.repository.pull_request.reviews;
-    Ok((conn.nodes, conn.page_info))
-}
-
 async fn fetch_review_threads(
     client: &GraphQLClient,
     repo: &RepoInfo,
@@ -501,14 +432,6 @@ async fn fetch_review_threads(
         };
     }
     Ok(threads)
-}
-
-async fn fetch_reviews(
-    client: &GraphQLClient,
-    repo: &RepoInfo,
-    number: u64,
-) -> Result<Vec<PullRequestReview>, VkError> {
-    paginate(|cursor| fetch_review_page(client, repo, number, cursor)).await
 }
 
 fn format_comment_diff(comment: &ReviewComment) -> Result<String, std::fmt::Error> {
@@ -685,51 +608,6 @@ fn summarize_files(threads: &[ReviewThread]) -> Vec<(String, usize)> {
         }
     }
     counts.into_iter().collect()
-}
-
-fn latest_reviews(reviews: Vec<PullRequestReview>) -> Vec<PullRequestReview> {
-    use std::collections::BTreeMap;
-    use std::collections::btree_map::Entry;
-
-    let mut map: BTreeMap<String, PullRequestReview> = BTreeMap::new();
-    for r in reviews {
-        let login = r
-            .author
-            .as_ref()
-            .map(|u| u.login.clone())
-            .unwrap_or_default();
-        match map.entry(login) {
-            Entry::Vacant(e) => {
-                e.insert(r);
-            }
-            Entry::Occupied(mut e) => {
-                if r.submitted_at > e.get().submitted_at {
-                    e.insert(r);
-                }
-            }
-        }
-    }
-    map.into_values().collect()
-}
-
-fn write_review<W: std::io::Write>(
-    mut out: W,
-    skin: &MadSkin,
-    review: &PullRequestReview,
-) -> anyhow::Result<()> {
-    let author = review.author.as_ref().map_or("", |u| u.login.as_str());
-    writeln!(out, "\u{1f4dd}  \x1b[1m{author}\x1b[0m {}:", review.state)?;
-    let _ = skin.write_text_on(&mut out, &review.body);
-    writeln!(out)?;
-    Ok(())
-}
-
-fn print_reviews(skin: &MadSkin, reviews: &[PullRequestReview]) {
-    for r in reviews {
-        if let Err(e) = write_review(std::io::stdout().lock(), skin, r) {
-            eprintln!("error printing review: {e}");
-        }
-    }
 }
 
 fn write_summary<W: std::io::Write>(
